refactor: make this a breaking change
