+++ firestore/src/main/java/com/firebase/ui/firestore/paging/FirestorePagingAdapter.java
@@ -7,80 +7,45 @@ import com.google.firebase.firestore.DocumentSnapshot;
 
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
-import androidx.arch.core.util.Function;
 import androidx.lifecycle.Lifecycle;
 import androidx.lifecycle.LifecycleObserver;
 import androidx.lifecycle.LiveData;
 import androidx.lifecycle.Observer;
 import androidx.lifecycle.OnLifecycleEvent;
-import androidx.lifecycle.Transformations;
-import androidx.paging.PagedList;
-import androidx.paging.PagedListAdapter;
+import androidx.paging.CombinedLoadStates;
+import androidx.paging.LoadState;
+import androidx.paging.PagingData;
+import androidx.paging.PagingDataAdapter;
 import androidx.recyclerview.widget.RecyclerView;
+import kotlin.Unit;
+import kotlin.jvm.functions.Function1;
 
 /**
- * This class is deprecated. Consider using {@link FirestorePagingDataAdapter} instead.
- *
  * Paginated RecyclerView Adapter for a Cloud Firestore query.
  *
  * Configured with {@link FirestorePagingOptions}.
  */
-@Deprecated
 public abstract class FirestorePagingAdapter<T, VH extends RecyclerView.ViewHolder>
-        extends PagedListAdapter<DocumentSnapshot, VH>
+        extends PagingDataAdapter<DocumentSnapshot, VH>
         implements LifecycleObserver {
 
-    private static final String TAG = "FirestorePagingAdapter";
-    /*
-        LiveData created via Transformation do not have a value until an Observer is attached.
-        We attach this empty observer so that our getValue() calls return non-null later.
-    */
-    private final Observer<FirestoreDataSource> mDataSourceObserver = new Observer<FirestoreDataSource>() {
-        @Override
-        public void onChanged(@Nullable FirestoreDataSource source) {
-
-        }
-    };
-    //Error observer to determine last occurred Error
-    private final Observer<Exception> mErrorObserver = new Observer<Exception>() {
-        @Override
-        public void onChanged(@Nullable Exception e) {
-            onError(e);
-        }
-    };
-    private final Observer<LoadingState> mStateObserver =
-            new Observer<LoadingState>() {
+    private final Observer<PagingData<DocumentSnapshot>> mDataObserver =
+            new Observer<PagingData<DocumentSnapshot>>() {
                 @Override
-                public void onChanged(@Nullable LoadingState state) {
-                    if (state == null) {
-                        return;
-                    }
-
-                    onLoadingStateChanged(state);
-                }
-            };
-    private final Observer<PagedList<DocumentSnapshot>> mDataObserver =
-            new Observer<PagedList<DocumentSnapshot>>() {
-                @Override
-                public void onChanged(@Nullable PagedList<DocumentSnapshot> snapshots) {
+                public void onChanged(@Nullable PagingData<DocumentSnapshot> snapshots) {
                     if (snapshots == null) {
                         return;
                     }
 
-                    submitList(snapshots);
+                    submitData(mOptions.getOwner().getLifecycle(), snapshots);
                 }
             };
     private FirestorePagingOptions<T> mOptions;
     private SnapshotParser<T> mParser;
-    private LiveData<PagedList<DocumentSnapshot>> mSnapshots;
-    private LiveData<LoadingState> mLoadingState;
-    private LiveData<Exception> mException;
-    private LiveData<FirestoreDataSource> mDataSource;
+    private LiveData<PagingData<DocumentSnapshot>> mSnapshots;
 
     /**
-     * This class is deprecated. Consider using {@link FirestorePagingDataAdapter} instead.
-     *
-     * Construct a new FirestorePagingAdapter from the given {@link FirestorePagingOptions}.
+     * Construct a new FirestorePagingDataAdapter from the given {@link FirestorePagingOptions}.
      */
     public FirestorePagingAdapter(@NonNull FirestorePagingOptions<T> options) {
         super(options.getDiffCallback());
@@ -91,67 +56,16 @@ public abstract class FirestorePagingAdapter<T, VH extends RecyclerView.ViewHold
     }
 
     /**
-     * Initializes Snapshots and LiveData
+     * Initializes Snapshots and LoadListener
      */
     private void init() {
-        mSnapshots = mOptions.getData();
-
-        mLoadingState = Transformations.switchMap(mSnapshots,
-                new Function<PagedList<DocumentSnapshot>, LiveData<LoadingState>>() {
-                    @Override
-                    public LiveData<LoadingState> apply(PagedList<DocumentSnapshot> input) {
-                        FirestoreDataSource dataSource = (FirestoreDataSource) input.getDataSource();
-                        return dataSource.getLoadingState();
-                    }
-                });
-
-        mDataSource = Transformations.map(mSnapshots,
-                new Function<PagedList<DocumentSnapshot>, FirestoreDataSource>() {
-                    @Override
-                    public FirestoreDataSource apply(PagedList<DocumentSnapshot> input) {
-                        return (FirestoreDataSource) input.getDataSource();
-                    }
-                });
-
-        mException = Transformations.switchMap(mSnapshots,
-                new Function<PagedList<DocumentSnapshot>, LiveData<Exception>>() {
-                    @Override
-                    public LiveData<Exception> apply(PagedList<DocumentSnapshot> input) {
-                        FirestoreDataSource dataSource = (FirestoreDataSource) input.getDataSource();
-                        return dataSource.getLastError();
-                    }
-                });
+        mSnapshots = mOptions.getPagingData();
 
         mParser = mOptions.getParser();
 
         if (mOptions.getOwner() != null) {
             mOptions.getOwner().getLifecycle().addObserver(this);
         }
-    }
-
-    /**
-     * If {@link #onLoadingStateChanged(LoadingState)} indicates error state, call this method to
-     * attempt to retry the most recent failure.
-     */
-    public void retry() {
-        FirestoreDataSource source = mDataSource.getValue();
-        if (source == null) {
-            Log.w(TAG, "Called retry() when FirestoreDataSource is null!");
-            return;
-        }
-        source.retry();
-    }
-
-    /**
-     * To attempt to refresh the list. It will reload the list from beginning.
-     */
-    public void refresh() {
-        FirestoreDataSource mFirebaseDataSource = mDataSource.getValue();
-        if (mFirebaseDataSource == null) {
-            Log.w(TAG, "Called refresh() when FirestoreDataSource is null!");
-            return;
-        }
-        mFirebaseDataSource.invalidate();
     }
 
     /**
@@ -182,9 +96,6 @@ public abstract class FirestorePagingAdapter<T, VH extends RecyclerView.ViewHold
     @OnLifecycleEvent(Lifecycle.Event.ON_START)
     public void startListening() {
         mSnapshots.observeForever(mDataObserver);
-        mLoadingState.observeForever(mStateObserver);
-        mDataSource.observeForever(mDataSourceObserver);
-        mException.observeForever(mErrorObserver);
     }
 
     /**
@@ -194,9 +105,6 @@ public abstract class FirestorePagingAdapter<T, VH extends RecyclerView.ViewHold
     @OnLifecycleEvent(Lifecycle.Event.ON_STOP)
     public void stopListening() {
         mSnapshots.removeObserver(mDataObserver);
-        mLoadingState.removeObserver(mStateObserver);
-        mDataSource.removeObserver(mDataSourceObserver);
-        mException.removeObserver(mErrorObserver);
     }
 
     @Override
@@ -210,23 +118,4 @@ public abstract class FirestorePagingAdapter<T, VH extends RecyclerView.ViewHold
      * @see #onBindViewHolder(RecyclerView.ViewHolder, int)
      */
     protected abstract void onBindViewHolder(@NonNull VH holder, int position, @NonNull T model);
-
-    /**
-     * Called whenever the loading state of the adapter changes.
-     * <p>
-     * When the state is {@link LoadingState#ERROR} the adapter will stop loading any data unless
-     * {@link #retry()} is called.
-     */
-    protected void onLoadingStateChanged(@NonNull LoadingState state) {
-        // For overriding
-    }
-
-    /**
-     * Called whenever the {@link Exception} is caught.
-     * <p>
-     * When {@link Exception} is caught the adapter will stop loading any data
-     */
-    protected void onError(@NonNull Exception e) {
-        Log.w(TAG, "onError", e);
-    }
 }
