+++ firestore/src/main/java/com/firebase/ui/firestore/paging/FirestorePagingOptions.java
@@ -10,8 +10,6 @@ import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import androidx.lifecycle.LifecycleOwner;
 import androidx.lifecycle.LiveData;
-import androidx.paging.LivePagedListBuilder;
-import androidx.paging.PagedList;
 import androidx.paging.Pager;
 import androidx.paging.PagingConfig;
 import androidx.paging.PagingData;
@@ -24,7 +22,7 @@ import static com.firebase.ui.common.Preconditions.assertNull;
 
 /**
  * Options to configure an {@link FirestorePagingAdapter}.
- *
+ * <p>
  * Use {@link Builder} to create a new instance.
  */
 public final class FirestorePagingOptions<T> {
@@ -32,27 +30,19 @@ public final class FirestorePagingOptions<T> {
     private static final String ERR_DATA_SET = "Data already set. " +
             "Call only one of setPagingData() or setQuery()";
 
-    private final LiveData<PagedList<DocumentSnapshot>> mData;
     private final LiveData<PagingData<DocumentSnapshot>> mPagingData;
     private final SnapshotParser<T> mParser;
     private final DiffUtil.ItemCallback<DocumentSnapshot> mDiffCallback;
     private final LifecycleOwner mOwner;
 
-    private FirestorePagingOptions(@NonNull LiveData<PagedList<DocumentSnapshot>> data,
-                                   @NonNull LiveData<PagingData<DocumentSnapshot>> pagingData,
+    private FirestorePagingOptions(@NonNull LiveData<PagingData<DocumentSnapshot>> pagingData,
                                    @NonNull SnapshotParser<T> parser,
                                    @NonNull DiffUtil.ItemCallback<DocumentSnapshot> diffCallback,
                                    @Nullable LifecycleOwner owner) {
-        mData = data;
         mPagingData = pagingData;
         mParser = parser;
         mDiffCallback = diffCallback;
         mOwner = owner;
-    }
-
-    @NonNull
-    public LiveData<PagedList<DocumentSnapshot>> getData() {
-        return mData;
     }
 
     @NonNull
@@ -80,60 +70,21 @@ public final class FirestorePagingOptions<T> {
      */
     public static final class Builder<T> {
 
-        private LiveData<PagedList<DocumentSnapshot>> mData;
         private LiveData<PagingData<DocumentSnapshot>> mPagingData;
         private SnapshotParser<T> mParser;
         private LifecycleOwner mOwner;
         private DiffUtil.ItemCallback<DocumentSnapshot> mDiffCallback;
 
         /**
-         *
-         * This method has been deprecated. Use {@link #setPagingData(LiveData, Class)}
-         * instead.
-         *
-         * Directly set data using and parse with a {@link ClassSnapshotParser} based on
-         * the given class.
-         * <p>
-         * Do not call this method after calling {@code setQuery}.
-         */
-        @Deprecated
-        @NonNull
-        public Builder<T> setData(@NonNull LiveData<PagedList<DocumentSnapshot>> data,
-                                              @NonNull Class<T> modelClass) {
-
-         return setData(data, new ClassSnapshotParser<>(modelClass));
-        }
-
-        /**
-         * Directly set data using and parse with a {@link ClassSnapshotParser} based on
-         * the given class.
+         * Directly set data using and parse with a {@link ClassSnapshotParser} based on the given
+         * class.
          * <p>
          * Do not call this method after calling {@code setQuery}.
          */
         @NonNull
         public Builder<T> setPagingData(@NonNull LiveData<PagingData<DocumentSnapshot>> data,
-                                  @NonNull Class<T> modelClass) {
+                                        @NonNull Class<T> modelClass) {
             return setPagingData(data, new ClassSnapshotParser<>(modelClass));
-        }
-
-        /**
-         *
-         * This method has been deprecated. Use {@link #setPagingData(LiveData, SnapshotParser)}
-         * instead.
-         *
-         * Directly set data and parse with a custom {@link SnapshotParser}.
-         * <p>
-         * Do not call this method after calling {@code setQuery}.
-         */
-        @Deprecated
-        @NonNull
-        public Builder<T> setData(@NonNull LiveData<PagedList<DocumentSnapshot>> data,
-                                  @NonNull SnapshotParser<T> parser) {
-            assertNull(mData, ERR_DATA_SET);
-
-            mData = data;
-            mParser = parser;
-            return this;
         }
 
         /**
@@ -143,7 +94,7 @@ public final class FirestorePagingOptions<T> {
          */
         @NonNull
         public Builder<T> setPagingData(@NonNull LiveData<PagingData<DocumentSnapshot>> pagingData,
-                                  @NonNull SnapshotParser<T> parser) {
+                                        @NonNull SnapshotParser<T> parser) {
             assertNull(mPagingData, ERR_DATA_SET);
 
             mPagingData = pagingData;
@@ -152,27 +103,9 @@ public final class FirestorePagingOptions<T> {
         }
 
         /**
-         * This method has been deprecated. Use {@link #setQuery(Query, PagingConfig, Class)}
-         * instead.
-         *
-         * Sets the query using {@link Source#DEFAULT} and a {@link ClassSnapshotParser} based
-         * on the given Class.
-         *
-         * See {@link #setQuery(Query, Source, PagedList.Config, SnapshotParser)}.
-         */
-        @Deprecated
-        @NonNull
-        public Builder<T> setQuery(@NonNull Query query,
-                                   @NonNull PagedList.Config config,
-                                   @NonNull Class<T> modelClass) {
-            return setQuery(query, Source.DEFAULT, config, modelClass);
-        }
-
-        /**
-         *
-         * Sets the query using {@link Source#DEFAULT} and a {@link ClassSnapshotParser} based
-         * on the given Class.
-         *
+         * Sets the query using {@link Source#DEFAULT} and a {@link ClassSnapshotParser} based on
+         * the given Class.
+         * <p>
          * See {@link #setQuery(Query, Source, PagingConfig, SnapshotParser)}.
          */
         @NonNull
@@ -183,25 +116,8 @@ public final class FirestorePagingOptions<T> {
         }
 
         /**
-         * This method has been deprecated.
-         * Use {@link #setQuery(Query, PagingConfig, SnapshotParser)} instead.
-         *
          * Sets the query using {@link Source#DEFAULT} and a custom {@link SnapshotParser}.
-         *
-         * See {@link #setQuery(Query, Source, PagedList.Config, SnapshotParser)}.
-         */
-        @Deprecated
-        @NonNull
-        public Builder<T> setQuery(@NonNull Query query,
-                                   @NonNull PagedList.Config config,
-                                   @NonNull SnapshotParser<T> parser) {
-            return setQuery(query, Source.DEFAULT, config, parser);
-        }
-
-        /**
-         *
-         * Sets the query using {@link Source#DEFAULT} and a custom {@link SnapshotParser}.
-         *
+         * <p>
          * See {@link #setQuery(Query, Source, PagingConfig, SnapshotParser)}.
          */
         @NonNull
@@ -212,28 +128,9 @@ public final class FirestorePagingOptions<T> {
         }
 
         /**
-         * This method has been deprecated.
-         * Use {@link #setQuery(Query, Source, PagingConfig, Class)} instead.
-         *
-         * Sets the query using a custom {@link Source} and a {@link ClassSnapshotParser} based
-         * on the given class.
-         *
-         * See {@link #setQuery(Query, Source, PagedList.Config, SnapshotParser)}.
-         */
-        @Deprecated
-        @NonNull
-        public Builder<T> setQuery(@NonNull Query query,
-                                   @NonNull Source source,
-                                   @NonNull PagedList.Config config,
-                                   @NonNull Class<T> modelClass) {
-            return setQuery(query, source, config, new ClassSnapshotParser<>(modelClass));
-        }
-
-        /**
-         *
-         * Sets the query using a custom {@link Source} and a {@link ClassSnapshotParser} based
-         * on the given class.
-         *
+         * Sets the query using a custom {@link Source} and a {@link ClassSnapshotParser} based on
+         * the given class.
+         * <p>
          * See {@link #setQuery(Query, Source, PagingConfig, SnapshotParser)}.
          */
         @NonNull
@@ -245,41 +142,10 @@ public final class FirestorePagingOptions<T> {
         }
 
         /**
-         * This method has been deprecated.
-         * Use {@link #setQuery(Query, Source, PagingConfig, SnapshotParser)} instead.
-         *
          * Sets the Firestore query to paginate.
          *
-         * @param query the Firestore query. This query should only contain where() and
-         *              orderBy() clauses. Any limit() or pagination clauses will cause errors.
-         * @param source the data source to use for query data.
-         * @param config paging configuration, passed directly to the support paging library.
-         * @param parser the {@link SnapshotParser} to parse {@link DocumentSnapshot} into model
-         *               objects.
-         * @return this, for chaining.
-         */
-        @Deprecated
-        @NonNull
-        public Builder<T> setQuery(@NonNull Query query,
-                                   @NonNull Source source,
-                                   @NonNull PagedList.Config config,
-                                   @NonNull SnapshotParser<T> parser) {
-            assertNull(mData, ERR_DATA_SET);
-
-            // Build paged list
-            FirestoreDataSource.Factory factory = new FirestoreDataSource.Factory(query, source);
-            mData = new LivePagedListBuilder<>(factory, config).build();
-
-            mParser = parser;
-            return this;
-        }
-
-        /**
-         *
-         * Sets the Firestore query to paginate.
-         *
-         * @param query the Firestore query. This query should only contain where() and
-         *              orderBy() clauses. Any limit() or pagination clauses will cause errors.
+         * @param query  the Firestore query. This query should only contain where() and orderBy()
+         *               clauses. Any limit() or pagination clauses will cause errors.
          * @param source the data source to use for query data.
          * @param config paging configuration, passed directly to the support paging library.
          * @param parser the {@link SnapshotParser} to parse {@link DocumentSnapshot} into model
@@ -291,18 +157,12 @@ public final class FirestorePagingOptions<T> {
                                    @NonNull final Source source,
                                    @NonNull PagingConfig config,
                                    @NonNull SnapshotParser<T> parser) {
-            assertNull(mData, ERR_DATA_SET);
             assertNull(mPagingData, ERR_DATA_SET);
 
             mParser = parser;
 
-            // Paging 2 support
-            PagedList.Config oldConfig = toOldConfig(config);
-            final FirestoreDataSource.Factory factory = new FirestoreDataSource.Factory(query, source);
-            mData = new LivePagedListBuilder<PageKey, DocumentSnapshot>(factory, oldConfig).build();
-
-
-            // For Paging 3 support
+            final FirestoreDataSource.Factory factory =
+                    new FirestoreDataSource.Factory(query, source);
             final Pager<PageKey, DocumentSnapshot> pager = new Pager<>(config,
                     new Function0<PagingSource<PageKey, DocumentSnapshot>>() {
                         @Override
@@ -311,18 +171,16 @@ public final class FirestorePagingOptions<T> {
                         }
                     });
 
-
             mPagingData = PagingLiveData.cachedIn(PagingLiveData.getLiveData(pager),
                     mOwner.getLifecycle());
             return this;
         }
 
 
-
         /**
-         * Sets an optional custom {@link DiffUtil.ItemCallback} to compare
-         * {@link DocumentSnapshot} objects.
-         *
+         * Sets an optional custom {@link DiffUtil.ItemCallback} to compare {@link DocumentSnapshot}
+         * objects.
+         * <p>
          * The default implementation is {@link DefaultSnapshotDiffCallback}.
          *
          * @return this, for chaining.
@@ -334,9 +192,9 @@ public final class FirestorePagingOptions<T> {
         }
 
         /**
-         * Sets an optional {@link LifecycleOwner} to control the lifecycle of the adapter. Otherwise,
-         * you must manually call {@link FirestorePagingAdapter#startListening()}
-         * and {@link FirestorePagingAdapter#stopListening()}.
+         * Sets an optional {@link LifecycleOwner} to control the lifecycle of the adapter.
+         * Otherwise, you must manually call {@link FirestorePagingAdapter#startListening()} and
+         * {@link FirestorePagingAdapter#stopListening()}.
          *
          * @return this, for chaining.
          */
@@ -351,7 +209,7 @@ public final class FirestorePagingOptions<T> {
          */
         @NonNull
         public FirestorePagingOptions<T> build() {
-            if ((mData == null && mPagingData == null) || mParser == null) {
+            if (mPagingData == null || mParser == null) {
                 throw new IllegalStateException("Must call setQuery() or setPagingData()" +
                         " before calling build().");
             }
@@ -360,25 +218,8 @@ public final class FirestorePagingOptions<T> {
                 mDiffCallback = new DefaultSnapshotDiffCallback<T>(mParser);
             }
 
-            return new FirestorePagingOptions<>(mData, mPagingData, mParser, mDiffCallback, mOwner);
+            return new FirestorePagingOptions<>(mPagingData, mParser, mDiffCallback, mOwner);
         }
-
-        /**
-         * Workaround to support the new PagingConfig class
-         * This should be removed once we fully migrate to Paging 3
-         * @param config the new PagingConfig
-         * @return the old PagedList.Config
-         */
-        private PagedList.Config toOldConfig(@NonNull PagingConfig config) {
-            return new PagedList.Config.Builder()
-                    .setEnablePlaceholders(config.enablePlaceholders)
-                    .setInitialLoadSizeHint(config.initialLoadSize)
-                    .setMaxSize(config.maxSize)
-                    .setPrefetchDistance(config.prefetchDistance)
-                    .setPageSize(config.pageSize)
-                    .build();
-        }
-
     }
 
 }
